requirements.txt
llama-index-llms-groq
llama-index
pinecone-client
openai
sentence-transformers
numpy
python-dotenv
pydantic
pytest
flask
gunicorn